requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
pyfiglet==1.0.4
python-dotenv==1.2.1
requests==2.32.5
tabulate==0.9.0
urllib3==2.5.0
